+++ src/index.ts
@@ -1,3 +1,11 @@
+import { registerPlugin } from '@capacitor/core';
+
+import type { OAuth2ClientPlugin } from './definitions';
+
+const OAuth2Client = registerPlugin<OAuth2ClientPlugin>('OAuth2Client', {
+    web: () => import('./web').then(m => new m.OAuth2ClientPluginWeb()),
+    // electron: () => ("./electron").then(m => new m.OAuth2ClientPluginElectron())
+});
+
 export * from './definitions';
-export * from './web-utils';
-export * from './web';
+export { OAuth2Client };